♻️ refactor: home 스타일 파일 컴포넌트 폴더로 코로케이션
