+++ src/components/home/guideCards.styles.ts
